renamed section components
